pyproject.toml
[project]
name = "datastarjinja"
version = "0.1.0"
description = "FastAPI project with DataStar and Jinja templates"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datastar-py>=0.4.3",
    "duckdb>=1.2.1",
    "fastapi-tailwind>=2.0.2a1",
    "fastapi[standard]>=0.115.11",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "jinja2>=3.1.3",
    "python-multipart>=0.0.9",
    "uuid>=1.30",
    "python-dateutil>=2.9.0",
    "sse-starlette>=1.6.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.optional-dependencies]
dev = [
    "black>=24.3.0",
    "pytest>=8.0.2",
]

[tool.ruff]
line-length = 100

[tool.black]
line-length = 100

[tool.pytest]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src"]

